server_assistant/requirements.txt
python-telegram-bot==20.7
torch==2.2.0
torchaudio==2.2.0
transformers==4.36.2
speechrecognition==3.10.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0  # Для работы с API
pytest==8.3.4
pytest-mock==3.14.0
gTTS==2.5.0
soundfile==0.12.1
sentencepiece==0.1.99
accelerate==0.26.1